ch4/ofi: code refactoring
Uninline non-performance-critical functions and move them to .c files.

Other making the conversion, two files are added for build correctness.

1. A ofi_noinline.h contains the declarations for all uninlined
functions and the conversion macro for NETMOD_INLINE build.

2. A netmod_impl.c for wrapper functions that is nolonger inlined.
