+++ src/mpid/ch4/netmod/Makefile.mk
@@ -16,6 +16,7 @@ AM_CPPFLAGS += -I$(top_srcdir)/src/mpid/ch4/netmod/include
 noinst_HEADERS += src/mpid/ch4/netmod/include/netmod.h
 noinst_HEADERS += src/mpid/ch4/netmod/include/netmod_impl.h
 
+include $(top_srcdir)/src/mpid/ch4/netmod/src/Makefile.mk
 include $(top_srcdir)/src/mpid/ch4/netmod/ofi/Makefile.mk
 include $(top_srcdir)/src/mpid/ch4/netmod/ucx/Makefile.mk
 include $(top_srcdir)/src/mpid/ch4/netmod/portals4/Makefile.mk
