+++ src/mpid/ch4/netmod/ofi/func_table.c
@@ -18,29 +18,29 @@ MPL_SUPPRESS_OSX_HAS_NO_SYMBOLS_WARNING;
 #include <mpidimpl.h>
 #include "netmod_inline.h"
 MPIDI_NM_funcs_t MPIDI_NM_ofi_funcs = {
-    .mpi_init = MPIDI_NM_mpi_init_hook,
-    .mpi_finalize = MPIDI_NM_mpi_finalize_hook,
+    .mpi_init = MPIDI_OFI_mpi_init_hook,
+    .mpi_finalize = MPIDI_OFI_mpi_finalize_hook,
     .progress = MPIDI_NM_progress,
-    .mpi_comm_connect = MPIDI_NM_mpi_comm_connect,
-    .mpi_comm_disconnect = MPIDI_NM_mpi_comm_disconnect,
-    .mpi_open_port = MPIDI_NM_mpi_open_port,
-    .mpi_close_port = MPIDI_NM_mpi_close_port,
-    .mpi_comm_accept = MPIDI_NM_mpi_comm_accept,
+    .mpi_comm_connect = MPIDI_OFI_mpi_comm_connect,
+    .mpi_comm_disconnect = MPIDI_OFI_mpi_comm_disconnect,
+    .mpi_open_port = MPIDI_OFI_mpi_open_port,
+    .mpi_close_port = MPIDI_OFI_mpi_close_port,
+    .mpi_comm_accept = MPIDI_OFI_mpi_comm_accept,
     /* Routines that handle addressing */
     .comm_get_lpid = MPIDI_NM_comm_get_lpid,
-    .get_local_upids = MPIDI_NM_get_local_upids,
-    .upids_to_lupids = MPIDI_NM_upids_to_lupids,
-    .create_intercomm_from_lpids = MPIDI_NM_create_intercomm_from_lpids,
-    .mpi_comm_create_hook = MPIDI_NM_mpi_comm_create_hook,
-    .mpi_comm_free_hook = MPIDI_NM_mpi_comm_free_hook,
+    .get_local_upids = MPIDI_OFI_get_local_upids,
+    .upids_to_lupids = MPIDI_OFI_upids_to_lupids,
+    .create_intercomm_from_lpids = MPIDI_OFI_create_intercomm_from_lpids,
+    .mpi_comm_create_hook = MPIDI_OFI_mpi_comm_create_hook,
+    .mpi_comm_free_hook = MPIDI_OFI_mpi_comm_free_hook,
     /* Window initialization/cleanup routines */
-    .mpi_win_create_hook = MPIDI_NM_mpi_win_create_hook,
-    .mpi_win_allocate_hook = MPIDI_NM_mpi_win_allocate_hook,
-    .mpi_win_allocate_shared_hook = MPIDI_NM_mpi_win_allocate_shared_hook,
-    .mpi_win_create_dynamic_hook = MPIDI_NM_mpi_win_create_dynamic_hook,
-    .mpi_win_attach_hook = MPIDI_NM_mpi_win_attach_hook,
-    .mpi_win_detach_hook = MPIDI_NM_mpi_win_detach_hook,
-    .mpi_win_free_hook = MPIDI_NM_mpi_win_free_hook,
+    .mpi_win_create_hook = MPIDI_OFI_mpi_win_create_hook,
+    .mpi_win_allocate_hook = MPIDI_OFI_mpi_win_allocate_hook,
+    .mpi_win_allocate_shared_hook = MPIDI_OFI_mpi_win_allocate_shared_hook,
+    .mpi_win_create_dynamic_hook = MPIDI_OFI_mpi_win_create_dynamic_hook,
+    .mpi_win_attach_hook = MPIDI_OFI_mpi_win_attach_hook,
+    .mpi_win_detach_hook = MPIDI_OFI_mpi_win_detach_hook,
+    .mpi_win_free_hook = MPIDI_OFI_mpi_win_free_hook,
     /* RMA synchronization routines */
     .rma_win_cmpl_hook = MPIDI_NM_rma_win_cmpl_hook,
     .rma_win_local_cmpl_hook = MPIDI_NM_rma_win_local_cmpl_hook,
@@ -75,8 +75,8 @@ MPIDI_NM_native_funcs_t MPIDI_NM_native_ofi_funcs = {
     .mpi_irecv = MPIDI_NM_mpi_irecv,
     .mpi_imrecv = MPIDI_NM_mpi_imrecv,
     .mpi_cancel_recv = MPIDI_NM_mpi_cancel_recv,
-    .mpi_alloc_mem = MPIDI_NM_mpi_alloc_mem,
-    .mpi_free_mem = MPIDI_NM_mpi_free_mem,
+    .mpi_alloc_mem = MPIDI_OFI_mpi_alloc_mem,
+    .mpi_free_mem = MPIDI_OFI_mpi_free_mem,
     .mpi_improbe = MPIDI_NM_mpi_improbe,
     .mpi_iprobe = MPIDI_NM_mpi_iprobe,
     .mpi_win_set_info = MPIDI_NM_mpi_win_set_info,
@@ -91,24 +91,24 @@ MPIDI_NM_native_funcs_t MPIDI_NM_native_ofi_funcs = {
     .mpi_win_unlock = MPIDI_NM_mpi_win_unlock,
     .mpi_win_get_info = MPIDI_NM_mpi_win_get_info,
     .mpi_get = MPIDI_NM_mpi_get,
-    .mpi_win_free = MPIDI_NM_mpi_win_free,
+    .mpi_win_free = MPIDI_OFI_mpi_win_free,
     .mpi_win_fence = MPIDI_NM_mpi_win_fence,
-    .mpi_win_create = MPIDI_NM_mpi_win_create,
+    .mpi_win_create = MPIDI_OFI_mpi_win_create,
     .mpi_accumulate = MPIDI_NM_mpi_accumulate,
-    .mpi_win_attach = MPIDI_NM_mpi_win_attach,
-    .mpi_win_allocate_shared = MPIDI_NM_mpi_win_allocate_shared,
+    .mpi_win_attach = MPIDI_OFI_mpi_win_attach,
+    .mpi_win_allocate_shared = MPIDI_OFI_mpi_win_allocate_shared,
     .mpi_rput = MPIDI_NM_mpi_rput,
     .mpi_win_flush_local = MPIDI_NM_mpi_win_flush_local,
-    .mpi_win_detach = MPIDI_NM_mpi_win_detach,
+    .mpi_win_detach = MPIDI_OFI_mpi_win_detach,
     .mpi_compare_and_swap = MPIDI_NM_mpi_compare_and_swap,
     .mpi_raccumulate = MPIDI_NM_mpi_raccumulate,
     .mpi_rget_accumulate = MPIDI_NM_mpi_rget_accumulate,
     .mpi_fetch_and_op = MPIDI_NM_mpi_fetch_and_op,
-    .mpi_win_allocate = MPIDI_NM_mpi_win_allocate,
+    .mpi_win_allocate = MPIDI_OFI_mpi_win_allocate,
     .mpi_win_flush = MPIDI_NM_mpi_win_flush,
     .mpi_win_flush_local_all = MPIDI_NM_mpi_win_flush_local_all,
     .mpi_win_unlock_all = MPIDI_NM_mpi_win_unlock_all,
-    .mpi_win_create_dynamic = MPIDI_NM_mpi_win_create_dynamic,
+    .mpi_win_create_dynamic = MPIDI_OFI_mpi_win_create_dynamic,
     .mpi_rget = MPIDI_NM_mpi_rget,
     .mpi_win_sync = MPIDI_NM_mpi_win_sync,
     .mpi_win_flush_all = MPIDI_NM_mpi_win_flush_all,
@@ -184,10 +184,10 @@ MPIDI_NM_native_funcs_t MPIDI_NM_native_ofi_funcs = {
     .mpi_ineighbor_alltoallv_sched = MPIDI_NM_mpi_ineighbor_alltoallv_sched,
     .mpi_ineighbor_alltoallw_sched = MPIDI_NM_mpi_ineighbor_alltoallw_sched,
     /* Datatype hooks */
-    .mpi_type_commit_hook = MPIDI_NM_mpi_type_commit_hook,
-    .mpi_type_free_hook = MPIDI_NM_mpi_type_free_hook,
+    .mpi_type_commit_hook = MPIDI_OFI_mpi_type_commit_hook,
+    .mpi_type_free_hook = MPIDI_OFI_mpi_type_free_hook,
     /* Op hooks */
-    .mpi_op_commit_hook = MPIDI_NM_mpi_op_commit_hook,
-    .mpi_op_free_hook = MPIDI_NM_mpi_op_free_hook,
+    .mpi_op_commit_hook = MPIDI_OFI_mpi_op_commit_hook,
+    .mpi_op_free_hook = MPIDI_OFI_mpi_op_free_hook,
 };
 #endif
